feed UI changes
